requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
bangla==0.0.4
blinker==1.9.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
colorama==0.4.6
comtypes==1.4.13
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.6
dateparser==1.1.8
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
einops==0.8.1
email-validator==2.3.0
encodec==0.1.1
exceptiongroup==1.3.0
fastapi==0.121.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
fuzzywuzzy==0.18.0
g2pkk==0.1.2
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
hangul-romanize==0.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
mutagen==1.47.0
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
omegaconf==2.3.0
openai==2.7.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
orjson==3.11.4
packaging==25.0
pandas==1.5.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
preshed==3.0.10
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pycparser==2.23
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
pypinyin==0.55.0
pypiwin32==223
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML==6.0.3
RapidFuzz==3.13.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.4
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.11.4
sentry-sdk==2.43.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.3
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.49.3
SudachiDict-core==20251022
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
trainer==0.0.36
transformers==4.57.1
TTS==0.22.0
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.0
yarl==1.22.0
zipp==3.23.0